♻️ partial class
